Rework standard pending op
Motivation
-----------
At present the type of the result returned by standard pending op
recv is specified at the point of recv. Whatever
creates the op should specify the type of the result as it already
knows what the type should be. This means that the creator can control
the result type and makes recv feel more ergonomic.

Changes
--------
Update standard pending op so that the result type is specified at
creation rather than at recv.