Apply deadline to connect/bootstrap ops
Motivation
----------
Bootstrap is only partially pipelined and is, in effect, a
synchronous process. We do not return a pending op from it. This
means that a higher up component cannot cancel any active requests
so we should apply the deadline to individual ops and connecting.

Changes
--------
Apply deadline to connect.
Rework bootstrap a bit so that only one operation is ever active
at once and apply deadline to each.